Move Perform Operation to it's own controller
Simplify the Ethereum Controller is used by other controllers, remove the Perform Operation method so that it's not unnecesarily included as part of the classes that extend it. The PerformOperation controller is now it's own class. Implemented ContainerInjectionInterface because it's a very simple class and doesn't necessarily need all the features of ControllerBase.

Also removed the deprecated redirect method, now using Drupal\Core\Url and Symfony\Component\HttpFoundation\RedirectResponse.